Fixed a bug in KeyExchanger. Arguments of the invocation for the timer now gets set correctly.
Changed FiSHy to adhere to the updated, not yet official, Colloquy plugin interface. This enables marking of decrypted messages, temporary overriding of encryption, and warning of encryption failures. If the user wanted encryption, but it fails for some reason, the message won't get sent. It will be displayed locally, though, but with a corresponding warning message.