One can now enable or disable encryption on a per room/nick basis. This influences if sent messages will be encrypted, and if received clear-text messages will be flagged with a warning.
A nick for which a keyx has been done will be flagged as encrypted until the program quits. Rooms/nicks for which a setkey has been done will be flagged as encrypted permanently. This can be changed with /enableEnc and /disableEnc

A message in an encrypted room can be send as clear-text via /+p