FiSHMisc: Added FiSHMisc as a central place to store common code. Right now, functions for casting JVChatViewController-conformant objects to JVDirectChatPanel (if possible), for getting the nickname/name of users/rooms, and for checking if a connection is IRC, are implemented. This saves a lot of duplicate code in FiSHController.
FiSHController:
Completly restructured processIncomingMessage:inView:. We now mark messages sent with encryption overridden as such.

We now check for each plugin method if we are working on a IRC connection, as this is the only type we support.


FiSHBlowfish:
Made encodeData: returning void, as the underlying dirt-functions always return success.

Project:
Set "Bundle Loader" to the Colloquy app. This way, we can at least use symbols from Colloquy. Symbols from Chat Core are still a no-go.